webui/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.1
requests==2.32.3
tenacity==8.4.2
pydantic==2.9.2
pandas==2.2.2
tqdm==4.66.5
PyYAML==6.0.1